修正補單BUG

+++ DocumentTemplateMaker.NET/Maker/Replace.cs
@@ -5,13 +5,16 @@ namespace DocumentTemplateMaker.NET;
 
 public class Replace
 {
-    public static void Maker(string tempFileName, string outputFileName, Parameter para)
+    public static void Maker(string tempFileName, string outputFileName, Parameter para, bool deleteAll = true)
     {
         // 讀檔案
         string text = System.IO.File.ReadAllText(tempFileName);
 
         // 砍擋
-        Helper.DeleteAll(".\\output\\");
+        if (deleteAll)
+        {
+            Helper.DeleteAll(".\\output\\");
+        }
 
         if (para.KeyWords.Length != para.ReplaceWords.Length)
         {
@@ -21,7 +24,18 @@ public class Replace
 
         for (int i = 0; i < para.KeyWords.Length; ++i)
         {
-            text = text.Replace(para.KeyWords[i], para.ReplaceWords[i]);
+            if (int.TryParse(para.KeyWords[i], out int n)
+                || float.TryParse(para.KeyWords[i], out float f)
+                || para.KeyWords[i].ToLower() == "false"
+                || para.KeyWords[i].ToLower() == "true")
+            {
+                text = text.Replace(para.KeyWords[i], para.ReplaceWords[i]);
+            }
+            else
+            {
+                text = text.Replace(para.KeyWords[i], '"' + para.ReplaceWords[i] + '"');
+            }
+            
         }
 
         // 寫檔案
